requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.2
Flask-APScheduler==1.13.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
typing-extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.17.0
